+++ src/net/kopeph/ld31/menu/TextBox.java
@@ -11,18 +11,18 @@ public class TextBox extends MenuWidget {
 	public String text;
 	public boolean xCenter, yCenter;
 
-	public TextBox(Font font, int x, int y, int w, int h) {
+	public TextBox(Font font, float x, float y, float w, float h) {
 		this(font, x, y, w, h, ""); //$NON-NLS-1$
 	}
 
-	public TextBox(Font font, int x, int y, int w, int h, String text) {
+	public TextBox(Font font, float x, float y, float w, float h, String text) {
 		super(x, y, w, h);
 		this.font = font;
 		this.text = text;
 	}
 
 	//This is sort of a hack for merging purposes
-	public TextBox(Font font, String text, int xPos, int yPos) {
+	public TextBox(Font font, String text, float xPos, float yPos) {
 		this(font, xPos, yPos, LD31.getContext().width, LD31.getContext().height, text);
 		xCenter = true;
 		yCenter = true;
